requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
plotly>=5.17.0
xlsxwriter>=3.1.0
reportlab>=4.0.0
kaleido==0.2.1  # v0.2.1.post1 unavailable; v1.x requires system Chrome
sqlalchemy>=2.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
python-dateutil>=2.8.0
authlib>=1.3.0
python-jose[cryptography]>=3.3.0
requests>=2.31.0